Move token_required to conf.py
